Added support for public tag sharing. Tags now have additional fields 'share', indicating shared status, and 'passcode', used to authenticate access.
Tags have a new resolver for updating the shared status.
Changing to 'public' generates a new passcode, while changing to 'none' removes the passcode.

Events, Contacts, and Content now have new unsecured resolvers that check for valid tag and passcode, and call the search resolver with that tag filter if valid.