Added location_id_string as a gettable/settable field on Events. Added a resolver 'findManyById' to Locations schema which takes in an array of IDs and gets the associated Locations.
This was in response to hydrating Locations on Events taking a very long time with millions of Locations per user.
I'm not sure why it was so slow, but not populating them on Events and instead getting them in a separate call is much faster overall.

Updated initial_searches.json with FontAwesome 5 icons.