Added write scope support to OAuth tokens. Made write scope cover read-scoped handlers. Moved Tracked and Uploaded Location bulk removal handlers to a new scope 'locations:admin' since those shouldn't be accessible to any user- generated OAuth app. Moved all OAuth app write-scoped handlers to 'oauthApps:admin' since, even though we don't hand out oauthApp scopes at all, those probably shouldn't ever be given out to a mere write-scoped token.
Better integrated OAuth apps with Provider/Connection schema.
OAuth apps now generate a Provider on creation.
When a user first completes the OAuth authorization workflow for
an OAuth app, a Connection is created for them to that app.
Further tokens generated for that app will be associated with that
original Connection.
App's Provider ID is shown on the app edit page, and the Connection ID
is returned alongside the token.
Deleting a Connection to an OAuth app removes it, all associated data,
and all associated tokens.

Updated Event bulkUpload handler to accept Locations as well.

Added eslint-disable for 'no-console' to logging statements that we
want in there.