Fixed some issues with OAuth app Provider/Connection logic:
Provider is no longer deleted when an OAuth app is deleted. This is so
that data generated by that app still has a valid Provider to refer to.

Provider's name is now the OAuth app's name. This is so that if the app
is deleted, the Provider still has a valid name reference.
Updating app's name updates the Provider's name as well as the name of
the Connections based on that Provider.
Provider queries no longer fetch the app since they already have a name.

Fixed some ingrained assumptions that a Provider will always have a map.

'runnable' field wasn't always being returned from Provider GQL calls,
especially subscription queries.