Added support for deauthorization and data deletion requests from Providers.
/auth/deauthorize-connection decodes an incoming payload and verifies its
signature. On success, it gets the user ID from the payload, looks up
what Connection that corresponds to, and then forces that Connection to
be unauthorized.

/auth/delete-connection decodes an incoming payload and verifies its
signature. On success, it gets the user ID from the payload, looks up
what Connection that corresponds to, and then deletes that Connection.
For Facebook and Instagram, it then creates a record of the data deletion
request with a confirmation code. The code and a URL containing that
code are sent back to the requesting server in JSON format.

A new schema structure, dataRemovalConfirmations, has been added.
Its one publicly-accessible handler takes in a confirmation code and
checks if that code corresponds to an existing confirmation and whether
the associated Connection still exists.

Added 'metadata' to the LifeScope Connection schema. This is copied from
the BitScoop Connection whenever a Connection is made or re-authorized.
This was necessary to handle the deauthorization and data deletion requests.